refactor: Use position type in current implementation

+++ src/main/scala/model/actors/ModelMessage.scala
@@ -5,6 +5,6 @@ import controller.Command
 import model.entities.{Bullet, Entity}
 
 trait ModelMessage
-case class Update(timeElapsed: Double, entities: List[Entity], replyTo: ActorRef[Command]) extends ModelMessage
+case class Update(timeElapsed: Float, entities: List[Entity], replyTo: ActorRef[Command]) extends ModelMessage
 case class Collision(entity: Entity, replyTo: ActorRef[Command]) extends ModelMessage
 case class Shoot(replyTo: ActorRef[Command]) extends ModelMessage
